2008-08-20  Doug Kwan  <dougkwan@google.com>
	* lto-section-out.c (produce_symtab_1): Fix a 32-bit bootstrap
	failure due to a missing cast.



git-svn-id: svn+ssh://gcc.gnu.org/svn/gcc/branches/lto@139345 138bc75d-0d04-0410-961f-82ee72b054a4

+++ gcc/lto-section-out.c
@@ -987,7 +987,7 @@ produce_symtab_1 (htab_t hash, struct lto_output_stream *stream,
 	}
 
       if (kind == LDPK_COMMON)
-	size = TREE_INT_CST_HIGH (DECL_SIZE (t)) << 32
+	size = ((uint64_t) TREE_INT_CST_HIGH (DECL_SIZE (t))) << 32
 	  | TREE_INT_CST_LOW (DECL_SIZE (t));
       else
 	size = 0;
